+++ eslint.config.js
@@ -1,8 +1,8 @@
-const js = require('@eslint/js');
-const prettier = require('eslint-config-prettier');
-const globals = require('globals');
+import js from '@eslint/js';
+import prettier from 'eslint-config-prettier';
+import globals from 'globals';
 
-module.exports = [
+export default [
     {
         ignores: ['node_modules/**'],
     },
@@ -11,8 +11,8 @@ module.exports = [
     {
         files: ['**/*.js'],
         languageOptions: {
-            ecmaVersion: 2022,
-            sourceType: 'commonjs',
+            ecmaVersion: 2025,
+            sourceType: 'module',
             globals: {
                 ...globals.node,
             },